Add Privacy level slider in to the Onboarding2PrivaciPanel[#963]
