+++ lib/ui/settings/SettingsNewPrivacyPanel.dart
@@ -30,6 +30,7 @@ import "package:illinois/service/Onboarding.dart";
 import "package:illinois/service/Storage.dart";
 import "package:illinois/ui/onboarding/OnboardingBackButton.dart";
 import "package:illinois/ui/widgets/HeaderBar.dart";
+import 'package:illinois/ui/widgets/PrivacySlider.dart';
 import "package:illinois/ui/widgets/RoundedButton.dart";
 import 'package:illinois/ui/widgets/ScalableWidgets.dart';
 import "package:illinois/ui/widgets/TabBarWidget.dart";
@@ -186,7 +187,7 @@ class SettingsNewPrivacyPanelState extends State<SettingsNewPrivacyPanel> implem
           top: false,
           child: Column(children: <Widget>[
             Container(height: 6,),
-            _PrivacyLevelSlider(
+            PrivacyLevelSlider(
               initialValue: _sliderValue,
               onValueChanged: (double value) {
                 if (value != _sliderValue) {
@@ -833,7 +834,7 @@ class _PrivacyEntryState extends State<_PrivacyEntry> with TickerProviderStateMi
           children: <Widget>[
             Container(
               padding: EdgeInsets.only(top: 4),
-              child: _PrivacyIcon(enabledIcon: iconRes, disabledIcon: iconResOff, minPrivacyLevel: minLevel, currentPrivacyLevel: widget.currentPrivacyLevel,)),
+              child: PrivacyIcon(enabledIcon: iconRes, disabledIcon: iconResOff, minPrivacyLevel: minLevel, currentPrivacyLevel: widget.currentPrivacyLevel,)),
             Container(width: 10,),
             Expanded(
              child: Column(
@@ -925,213 +926,6 @@ class _PrivacyEntryState extends State<_PrivacyEntry> with TickerProviderStateMi
         )
       ],
     );
-  }
-}
-
-class _PrivacyLevelSlider extends StatefulWidget {
-  final double initialValue;
-  final Function onValueChanged;
-
-  const _PrivacyLevelSlider({Key key, this.onValueChanged, this.initialValue}) : super(key: key);
-
-  @override
-  _PrivacyLevelSliderState createState() => _PrivacyLevelSliderState();
-}
-
-class _PrivacyLevelSliderState extends State<_PrivacyLevelSlider> {
-  double _discreteValue;
-  Color _mainColor = Styles().colors.white;
-  Color _trackColor = Styles().colors.fillColorPrimaryVariant;
-  Color _inactiveTrackColor = Styles().colors.surfaceAccent;
-
-  @override
-  void initState() {
-    super.initState();
-    _discreteValue = widget.initialValue;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    int _roundedValue = _discreteValue?.round();
-    final ThemeData theme = Theme.of(context);
-    return Container(
-        padding: EdgeInsets.symmetric(horizontal: 17),
-        color: Styles().colors.white,
-        child: Column(
-          children: <Widget>[
-            Stack(
-            children: <Widget>[
-            Positioned.fill(
-            child: Align(
-                alignment: Alignment.center,
-                child: Padding(
-                  padding: EdgeInsets.only(left: 10, right: 20),
-                  child: Container(
-                      height: 11,
-                      decoration: BoxDecoration(
-                        color: Styles().colors.fillColorPrimaryVariant,
-                        border: Border.all(color: Styles().colors.fillColorPrimaryVariant, width: 1),
-                        borderRadius: BorderRadius.circular(24.0),
-                      )),
-                )),
-              ),
-              SliderTheme(
-                  data: theme.sliderTheme.copyWith(
-                      activeTrackColor: _trackColor,
-                      inactiveTrackColor: _inactiveTrackColor,
-                      activeTickMarkColor: _mainColor,
-                      thumbColor: _mainColor,
-                      thumbShape: _CustomThumbShape(),
-                      tickMarkShape: _CustomTickMarkShape(),
-                      trackHeight: 10,
-                      inactiveTickMarkColor: _inactiveTrackColor,
-                      showValueIndicator: ShowValueIndicator.never,
-                      valueIndicatorTextStyle: TextStyle(fontSize: 20, fontFamily: Styles().fontFamilies.extraBold, color: Styles().colors.fillColorPrimary)),
-                  child: MergeSemantics(
-                      child: Semantics(
-                          label: Localization().getStringEx("panel.settings.new_privacy.privacy.button.set_privacy.slider.hint", "Privacy Level"),
-                          enabled: true,
-                          increasedValue: Localization().getStringEx("panel.settings.new_privacy.privacy.button.set_privacy.slider.increase", "increased to") +
-                              (_roundedValue + 1).toString(),
-                          decreasedValue: Localization().getStringEx("panel.settings.new_privacy.privacy.button.set_privacy.slider.decrease", "decreased to") +
-                              (_roundedValue - 1).toString(),
-                          child:
-                          Padding(
-                            padding: EdgeInsets.symmetric(horizontal: 2, vertical: 2), //Fix cut off circle
-                            child:
-                            Slider(
-                              value: _discreteValue,
-                              min: 1.0,
-                              max: 5.0,
-                              divisions: 5,
-                              semanticFormatterCallback: (double value) => value.round().toString(),
-                              label: "$_roundedValue",
-                              onChanged: (double value) {
-                                setState(() {
-                                  _discreteValue = value;
-                                  if(value>3.3 && value<4) { // remove the second 3rd division caused by {max == division}
-                                    _discreteValue = value = 4;
-                                  }
-                                  widget.onValueChanged(value);
-                                });
-                              },
-                            )
-              ))))
-              ]),
-              Container(
-                padding: EdgeInsets.symmetric(horizontal: 18, vertical: 11),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: <Widget>[
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 1,
-                      enabledIcon: "images/view-only-blue.png",
-                      disabledIcon: "images/view-only-blue.png",
-                    ),
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 2,
-                      enabledIcon: "images/location-sharing-blue.png",
-                      disabledIcon: "images/location-sharing-blue-off.png",
-                    ),
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 3,
-                      enabledIcon: "images/personalization-blue.png",
-                      disabledIcon: "images/personalization-blue-off.png",
-                    ),
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 4,
-                      enabledIcon: "images/notifications-blue.png",
-                      disabledIcon: "images/notifications-blue-off.png",
-                    ),
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 4,
-                      enabledIcon: "images/identiy-blue.png",
-                      disabledIcon: "images/identiy-blue-off.png",
-                    ),
-                    _PrivacyIcon(
-                      currentPrivacyLevel: _currentLevel,
-                      minPrivacyLevel: 5,
-                      enabledIcon: "images/share-blue.png",
-                      disabledIcon: "images/share-blue-off.png",
-                    ),
-                  ],
-                )
-              ),
-            Container(height: 5,)
-            ]
-        ));
-  }
-
-  int get _currentLevel{
-    return _discreteValue?.round();
-  }
-}
-
-class _PrivacyIcon extends StatelessWidget{
-  final int currentPrivacyLevel;
-  final int minPrivacyLevel;
-  final enabledIcon;
-  final disabledIcon;
-
-  const _PrivacyIcon({Key key, this.minPrivacyLevel = 1, this.enabledIcon, this.disabledIcon, this.currentPrivacyLevel}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Image.asset(currentPrivacyLevel>=minPrivacyLevel? enabledIcon : (disabledIcon??""), excludeFromSemantics: true,);
-  }
-
-}
-
-class _CustomThumbShape extends SliderComponentShape {
-  @override
-  Size getPreferredSize(bool isEnabled, bool isDiscrete) {
-    return const Size.fromRadius(16);
-  }
-
-  @override
-  void paint(PaintingContext context, Offset center, {Animation<double> activationAnimation, Animation<double> enableAnimation, bool isDiscrete, TextPainter labelPainter, RenderBox parentBox, SliderThemeData sliderTheme, TextDirection textDirection, double value, double textScaleFactor, Size sizeWithOverflow}) {
-    final Canvas canvas = context.canvas;
-    final ColorTween colorTween = ColorTween(
-      begin: sliderTheme.disabledThumbColor,
-      end: sliderTheme.activeTickMarkColor,
-    );
-
-    final ColorTween colorTween2 = ColorTween(
-      begin: Styles().colors.white,
-      end: Styles().colors.white,
-    );
-
-    final ColorTween colorTween3 = ColorTween(
-      begin: Styles().colors.fillColorSecondary,
-      end: Styles().colors.fillColorSecondary,
-    );
-    final ColorTween colorTween4 = ColorTween(
-      begin: Styles().colors.fillColorPrimary,
-      end: Styles().colors.fillColorPrimary,
-    );
-
-    canvas.drawCircle(center, 25, Paint()..color = colorTween4.evaluate(enableAnimation));
-    canvas.drawCircle(center, 23, Paint()..color = colorTween2.evaluate(enableAnimation));
-    canvas.drawCircle(center, 21, Paint()..color = colorTween3.evaluate(enableAnimation));
-    canvas.drawCircle(center, 19, Paint()..color = colorTween.evaluate(enableAnimation));
-    labelPainter.paint(canvas, center + Offset(-labelPainter.width / 2.0, -labelPainter.height / 2.0));
-  }
-}
-
-class _CustomTickMarkShape extends SliderTickMarkShape {
-  @override
-  Size getPreferredSize({SliderThemeData sliderTheme, bool isEnabled}) {
-    return Size.fromRadius(3);
-  }
-
-  @override
-  void paint(PaintingContext context, Offset center,
-      {RenderBox parentBox, SliderThemeData sliderTheme, Animation<double> enableAnimation, Offset thumbCenter, bool isEnabled, TextDirection textDirection}) {
-      //Don"t draw - we don"t want to show TickMarkers
   }
 }
